etcdserver: add ability to auto-promote learners to voters
It looks as though the etcd team was planning to implement
"auto promote" (see #10537). This commit attempts to lay the
groundwork for auto-promoting learners to voters by adding an
--auto-promote flag to the add member command. One reason
for having the ability to automatically promote learners to
voters is that it would enable implementing the Raft paper's
recommendation handle reconfiguration changes by first adding
new members as non-voters until they have caught up with the
leader.