etcdserver: add ability to auto-promote learners to voters
Enable new learners to be marked for automatic promotion to voters upon
catching up with the leader.

Adds the ability to supply one or more promotion rules when adding a new
learnere. Promotion rules govern if and when a learner may be promoted
to a new role. Currently only promotion to voter is supported, but the
APIs added in this commit are flexible and allow for
backwards-compatible introduction of additional promotion targets (e.g.
reader).